continue grinding through coreData!

+++ DiscGolfSearch/Models/BagDataModel.swift
@@ -9,13 +9,13 @@ import UIKit
 import SwiftData
 
 struct BagDataModel {
-    let id: Int
+    let id: Int64
     let bagHexColor: String
     let bagTitle: String
     let bagType: String
     let discs: [DiscGolfDisc]
     
-    init(id: Int, bagHexColor: String, bagTitle: String, bagType: String) {
+    init(id: Int64, bagHexColor: String, bagTitle: String, bagType: String) {
         self.id = id
         self.bagHexColor = bagHexColor
         self.bagTitle = bagTitle
